requirements.txt
# LangGraph
langgraph>=0.3.0
langchain-core>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# PDF Parsing
PyMuPDF>=1.24.0

# Knowledge Store
kuzu>=0.11.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
Pillow>=10.0.0
